chore: Move components and composables in app subfolder
