+++ src/runtime/server/tsconfig.json
@@ -1,3 +1,3 @@
 {
-  "extends": "../../../.nuxt/tsconfig.server.json",
+  "extends": "../../../.nuxt/tsconfig.server.json"
 }
